refactor(TestDatabasePopulatorUtils): 테스트용 DB 초기화 클래스
-  DatabasePopulatorUtils에 추가했던 메서드 삭제
